+++ app/src/main/java/com/matheus/planningapp/viewmodel/commitment/CommitmentFormMode.kt
@@ -4,11 +4,11 @@ import kotlinx.datetime.Instant
 
 sealed interface CommitmentFormMode {
     data class Create(
-        val calendarId: Int,
+        val calendarId: Long,
         val initialInstant: Instant
     ): CommitmentFormMode
 
     data class Edit(
-        val commitmentId: Int
+        val commitmentId: Long
     ): CommitmentFormMode
 }
